refactor(tel-ui): extract function, change var names
