+++ telemetry/packages/constants/src/index.ts
@@ -9,7 +9,7 @@ export {
   getStateType,
 } from './pods/states';
 export type { PodStateType, PodStateCategoryType } from './pods/states';
-export { MODES, MODE_INACTIVE_STATES } from './pods/modes';
+export { MODES, MODE_EXCLUDED_STATES } from './pods/modes';
 export type { ModeType } from './pods/modes';
 export { openMctObjectTypes } from './openmct/object-types/object-types';
